pyproject.toml
[project]
name = "web-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.9",
]
